requirements.txt
azure-core==1.17.0
azure-cosmos==4.2.0
Brotli==1.0.9
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dash==1.21.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.2
plotly==5.2.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
tenacity==8.0.1
urllib3==1.26.6
Werkzeug==2.0.1
